+++ src/types/keywordOption.ts
@@ -1,0 +1,4 @@
+export type KeywordOptions = {
+    keywordMap: Record<string, string[]>;
+    labelMap: Record<string, string>;
+};
