+++ src/utils/normalizer.ts
@@ -1,0 +1,3 @@
+export const normalizeEmptyToNull = (value: string | undefined): string | null => {
+    return value && value.trim() !== '' ? value : null;
+};
